+++ client/src/app/LinkedInLogin.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+export default class LinkedInLogin extends React.Component {
+    constructor(props) {
+        super(props)
+        this.state = {}
+    }
+    render() {
+      return ( 
+        <div>
+          <button>hi</button>
+        </div>
+      );
+    }
+};
+
