+++ client/src/images/coin.gif
