+++ client/src/sessions/SessionChoice.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Mutation } from 'react-apollo';
 import { Modal, Button } from 'react-bootstrap';
-import { UPDATE_SESSION } from '../../gql.js';
+import { UPDATE_SESSION } from '../apollo/gql.js';
 import { Link } from 'react-router-dom';
 
 export default class SessionChoice extends Component {
